Move FB logic to another file
